[UPDATE] #S08P12A609-37 JWT 헤더 사용자 인증

+++ back/src/main/java/com/ssafy/logit/controller/user/UserController.java
@@ -20,6 +20,9 @@ public class UserController {
 
     private static final String SUCCESS = "success";
     private static final String FAIL = "fail";
+    private static final String UNAUTHORIZED = "unauthorized";
+    private static final String DELETED = "deleted";
+    private static final String NONE = "none";
 
     @Autowired
     private UserService userService;
@@ -45,13 +48,14 @@ public class UserController {
         return new ResponseEntity<Map<String, Object>>(resultMap, HttpStatus.ACCEPTED);
     }
 
-    // refresh-token 확인 후 auth-token 재발급
+    // 토큰 재발급
     @PostMapping("/refresh")
     public ResponseEntity<Map<String, Object>> refreshToken(@RequestParam String email) {
         Map<String, Object> resultMap = new HashMap<>();
         String token = userService.getRefreshToken(email);
         jwtUtil.checkAndGetClaims(token);
 
+        // refresh-token 확인 후 auth-token 재발급
         if (token.equals(userService.getRefreshToken(email))) {
             String authToken = jwtUtil.createAuthToken(email);
             resultMap.put("jwt-auth-token", authToken);
@@ -64,17 +68,32 @@ public class UserController {
 
     // 로그아웃
     @PostMapping("/logout")
-    public ResponseEntity<String> logout(@RequestParam String email) {
-        log.debug("logout : {}", email);
-        userService.logout(email);
-        return new ResponseEntity<String>(SUCCESS, HttpStatus.ACCEPTED);
+    public ResponseEntity<String> logout(@RequestAttribute String email) {
+        try {
+            log.debug("logout : {}", email);
+            userService.logout(email);
+            return new ResponseEntity<String>(SUCCESS, HttpStatus.ACCEPTED);
+        } catch (Exception e) {
+            e.printStackTrace();
+            return new ResponseEntity<>(FAIL, HttpStatus.OK);
+        }
     }
 
     // 회원 삽입 or 수정
     @PostMapping
-    public ResponseEntity<String> saveUser(@RequestBody UserDto userDto) throws Exception {
+    public ResponseEntity<String> saveUser(@RequestBody UserDto userDto, @RequestAttribute String email) throws Exception {
         try {
-            userService.saveUser(userDto);
+            // 해당하는 회원 없을 경우 <회원 가입>, 있을 경우 <회원 정보 수정>
+            boolean regist = true;
+            if(userService.getUser(userDto.getEmail()) != null) {
+                regist = false;
+
+                // 토큰 사용자 인증
+                if(!userDto.getEmail().equals(email)) {
+                    return new ResponseEntity<>(UNAUTHORIZED, HttpStatus.UNAUTHORIZED);
+                }
+            }
+            userService.saveUser(userDto, regist);
             return new ResponseEntity<String>(SUCCESS, HttpStatus.OK);
         } catch (Exception e) {
             e.printStackTrace();
@@ -89,22 +108,27 @@ public class UserController {
     }
 
     // email로 회원 조회
-    @GetMapping("/{email}")
-    public ResponseEntity<UserDto> getUser(@PathVariable String email) throws Exception {
+    @GetMapping
+    public ResponseEntity<UserDto> getUser(@RequestAttribute String email) throws Exception {
         return new ResponseEntity<UserDto>(userService.getUser(email), HttpStatus.OK);
     }
 
     // 회원 삭제 (실제 삭제x, isDeleted 1로 업데이트)
     @PutMapping("/{id}")
-    public ResponseEntity<String> deleteUser(@PathVariable Long id) throws  Exception {
-        String result = userService.deleteUser(id);
+    public ResponseEntity<String> deleteUser(@PathVariable Long id, @RequestAttribute String email) throws  Exception {
         try {
-            if(result.equals("success")) {
-                return new ResponseEntity<String>(SUCCESS, HttpStatus.OK);
-            } else if(result.equals("already deleted")) {
-                return new ResponseEntity<String>("이미 삭제된 회원입니다.", HttpStatus.OK);
+            // 토큰 사용자 인증 후 회원 삭제
+            if(userService.getUser(id).getEmail().equals(email)) {
+                String result = userService.deleteUser(id);
+                if(result.equals(SUCCESS)) { // delete 성공
+                    return new ResponseEntity<String>(SUCCESS, HttpStatus.OK);
+                } else if(result.equals(DELETED)) { // 이미 delete 된 회원
+                    return new ResponseEntity<String>(DELETED, HttpStatus.OK);
+                } else { // 없는 회원
+                    return new ResponseEntity<String>(NONE, HttpStatus.OK);
+                }
             } else {
-                return new ResponseEntity<String>("없는 회원입니다.", HttpStatus.OK);
+                return new ResponseEntity<>(UNAUTHORIZED, HttpStatus.UNAUTHORIZED);
             }
         } catch (Exception e) {
             e.printStackTrace();
@@ -112,7 +136,7 @@ public class UserController {
         }
     }
 
-    // 회원 삭제 (실제 삭제)
+    // 회원 삭제 (실제 삭제) - 회원 이용 불가 !!
     @DeleteMapping("/{id}")
     public ResponseEntity<String> dropUser(@PathVariable Long id) throws Exception {
         boolean result = userService.dropUser(id);
@@ -120,7 +144,7 @@ public class UserController {
             if(result) {
                 return new ResponseEntity<>(SUCCESS, HttpStatus.OK);
             } else {
-                return new ResponseEntity<>("없는 회원입니다.", HttpStatus.OK);
+                return new ResponseEntity<>(NONE, HttpStatus.OK);
             }
         } catch (Exception e) {
             e.printStackTrace();
